When done with Mapping, refactor annotations
Refactor annotations to instance variables rather than getters. Do this to see if it's easier to read and understand. Because POJO's can be long, the importance of knowning where the annotations are connected to make sense to have at the top of the POJO above the instance variables so you don't have to go searching for them in the getters