Added util package
In util package added DateTimeUtil class so it's possible to grab the users system Time and Date. To be able to grab the users Date and Time with their time Zone I should probaly user ZonedDateTime instead of LocalDateTime. For now this works, but moving forward I should think about the users Time Zone and code it out accordinglly so I'm not mistaken on when accounts were created.